# store-application/src/main/java/dev/infinia/store/app/web/CatalogController.java
package dev.infinia.store.app.web;

import dev.infinia.store.app.service.CatalogService;
import dev.infinia.store.contract.api.CatalogDtos;
import dev.infinia.store.contract.type.Channel;
import dev.infinia.store.contract.type.ListingType;
import dev.infinia.store.domain.port.ListingQuery;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RestController;

/** Public catalog browsing — anonymous access allowed (design §10.2, §13.2). */
@RestController
@RequestMapping("/api/v1")
public class CatalogController {

    private final CatalogService catalog;

    public CatalogController(CatalogService catalog) {
        this.catalog = catalog;
    }

    @GetMapping("/catalog")
    public ResponseEntity<CatalogDtos.CatalogPageDto> browse(
            @RequestParam(required = false) String type,
            @RequestParam(required = false) String query,
            @RequestParam(required = false) String category,
            @RequestParam(required = false) String channel,
            @RequestParam(required = false) String hostVersion,
            @RequestParam(required = false) String os,
            @RequestParam(required = false) String arch,
            @RequestParam(required = false) String sort,
            @RequestParam(required = false, name = "featured") Boolean featured,
            @RequestParam(required = false) String cursor,
            @RequestParam(defaultValue = "24") int limit) {
        CatalogDtos.CatalogPageDto page = catalog.browse(new CatalogService.BrowseQuery(
                type == null || type.isBlank() ? null
                        : ListingType.valueOf(type.trim().toUpperCase()),
                query,
                category,
                channel == null ? null : Channel.valueOf(channel.trim().toUpperCase()),
                hostVersion,
                os,
                arch,
                sort == null ? null : ListingQuery.ListingSort.valueOf(sort.trim().toUpperCase()),
                featured,
                cursor,
                Math.min(Math.max(limit, 1), 100)));
        String etag = "\"catalog-" + Integer.toHexString(page.hashCode()) + "\"";
        // no-cache (revalidate-always) rather than max-age: a stale browser copy
        // would hide freshly relisted listings for the whole TTL after admin
        // delist/relist cycles. ETag revalidation keeps unchanged responses 304.
        // Vary: the bee-level gate makes identical URLs differ per Authorization.
        return ResponseEntity.ok().eTag(etag).header("Cache-Control", "no-cache")
                .header("Vary", "Authorization")
                .body(page);
    }
}