# store-contract/src/main/resources/contract/openapi.yaml
openapi: 3.1.0
info:
  title: Infinia Store Platform API
  version: 1.0.0
  summary: Catalog, publishing, review and delivery API for the Infinia / FengYu ecosystem.
  description: |
    Versioned REST API of the Infinia Store Platform (design §10).

    * Root path `/api/v1`; breaking changes ship as `/api/v2` — never via headers.
    * JSON is camelCase, timestamps are UTC RFC 3339, identifiers are UUIDv7/ULID.
    * Errors use `application/problem+json` with a stable `code`, `traceId` and
      localizable arguments.
    * Lists use cursor pagination; write endpoints accept `Idempotency-Key`;
      concurrent updates use `ETag/If-Match`.
  license:
    name: GPL-3.0
    url: https://www.gnu.org/licenses/gpl-3.0.txt
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
  /catalog:
    get:
      operationId: browseCatalog
      tags: [Catalog]
      summary: Browse the public catalog (anonymous access allowed)
      security:
        - {}
        - bearerAuth: []
      parameters:
        - { name: type, in: query, schema: { $ref: '#/components/schemas/ListingType' } }
        - { name: query, in: query, schema: { type: string, maxLength: 128 }, description: Free-text search over name, summary and tags }
        - { name: category, in: query, schema: { type: string } }
        - { name: channel, in: query, schema: { $ref: '#/components/schemas/Channel' } }
        - { name: hostVersion, in: query, schema: { type: string }, description: Only listings with a compatible release }
        - { name: os, in: query, schema: { $ref: '#/components/schemas/Platform' } }
        - { name: arch, in: query, schema: { $ref: '#/components/schemas/Arch' } }
        - { name: sort, in: query, schema: { type: string, enum: [relevance, recent, downloads, favorites], default: relevance } }
        - { name: featured, in: query, schema: { type: boolean }, description: Editorial featured shelf only }
        - { name: cursor, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 24 } }
      responses:
        '200':
          description: Catalog page
          headers:
            ETag: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CatalogPage' }
...
  /compat/fengyu/codex/catalog:
    get:
      operationId: fengyuCodexCatalog
      tags: [Compatibility]
      summary: Codex-consumable skill catalog pointing at native manifests
      security:
        - {}
      responses:
        '200':
          description: Codex catalog
          content:
            application/json:
              schema:
                type: object
                properties:
                  schemaVersion: { type: integer }
                  generator: { type: string }
                  skills:
                    type: array
                    items:
                      type: object
                      properties:
                        name: { type: string }
                        skillId: { type: string }
                        version: { type: string }
                        listing: { type: string }
                        installManifest: { type: string }
  /compat/fengyu/claude-marketplace.json:
    get:
      operationId: claudeEcosystemMarketplace
      tags: [Compatibility]
      summary: CLAUDE-ecosystem marketplace exporting skills and MCP templates as git sources
      description: |
        Each published SKILL/MCP listing is materialized as a small local git
        repository carrying a .claude-plugin/plugin.json (skills paths and an
        mcpServers block). Registering this URL as a CLAUDE-type source in the
        FengYu host installs skills and disabled MCP server definitions through
        the host's existing ecosystem installer; re-cloning picks up updates.
      security:
        - {}
      responses:
        '200':
          description: Claude marketplace document
          content:
            application/json:
              schema:
                type: object
                properties:
                  plugins:
                    type: array
                    items:
                      type: object
                      properties:
                        name: { type: string }
                        description: { type: string }
                        category: { type: string }
                        keywords: { type: array, items: { type: string } }
                        author:
                          type: object
                          properties:
                            name: { type: string }
                        source:
                          type: object
                          properties:
                            source: { type: string, enum: [url] }
                            url: { type: string, description: file:// clone URL }
  /compat/fengyu/skills-catalog:
    get:
      operationId: fengyuSkillsCatalog
      tags: [Compatibility]
      summary: SkillCatalogEntry-compatible skill catalog (fengyu.skills.catalog-url)
      security:
        - {}
      responses:
        '200':
          description: Legacy skill catalog entries
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id: { type: string }
                    name: { type: string }
                    description: { type: string }
                    version: { type: string }
                    author: { type: string }
                    downloadUrl: { type: string }
                    official: { type: boolean }
  /compat/fengyu/fengyu-releases/api/releases/latest:
    get:
      operationId: fengyuPortableReleaseMirror
      tags: [Compatibility]
      summary: GitHub-releases-compatible mirror for the Electron portable updater
      description: Consumed via the host's updateApiBase setting ({base}/fengyu-releases/api/releases/latest).
      security:
        - {}
      parameters:
        - { name: channel, in: query, schema: { type: string } }
      responses:
        '200':
          description: GitHub-compatible release object, or [] when no stable APP release exists
          content:
            application/json:
              schema: {}
  /compat/fengyu/catalog:
    get:
      operationId: fengyuCompatCatalog
      tags: [Compatibility]
      summary: MarketplaceCatalogEntry-compatible plugin catalog for the FengYu host store
      description: |
        Serves the JSON array contract the FengYu host's built-in FENGYU marketplace
        source type fetches, so the host lists, updates and downloads plugins with
        zero code changes. Each entry carries the latest published version and a
        direct HMAC-ticketed downloadUrl for the package artifact.
      security:
        - {}
      responses:
        '200':
          description: Legacy catalog entries
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/FengYuCatalogEntry' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
    ListingType:
      type: string
      enum: [APP, PLUGIN, SKILL, MCP, FLOW]
    Channel:
      type: string
      enum: [stable, beta, alpha, nightly, private]
    Platform:
      type: string
      enum: [windows, macos, linux, universal]
    Arch:
      type: string
      enum: [x64, arm64, universal]
...
    CatalogItem:
      type: object
      required: [coordinate, type, namespace, slug, name]
      properties:
        coordinate: { type: string, example: 'infinia://plugin/official/markdown' }
        type: { $ref: '#/components/schemas/ListingType' }
        namespace: { type: string }
        slug: { type: string }
        name: { type: string }
        summary: { type: string }
        category: { type: string }
        tags: { type: array, items: { type: string } }
        iconUrl: { type: string, format: uri }
        latestVersion: { type: string }
        channel: { $ref: '#/components/schemas/Channel' }
        downloads: { type: integer, format: int64 }
        publisherName: { type: string }
        updatedAt: { type: string, format: date-time }
        featured: { type: boolean, description: Editorial featuring (admin) }
        minBeeLevel: { type: integer, minimum: 0, maximum: 4, description: 'Infinia Level gate: 0 = public; 1..4 restricts view/download to accounts at or above that level' }
    CatalogPage:
      type: object
      required: [items]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/CatalogItem' } }
        nextCursor: { type: string, nullable: true }
        totalEstimate: { type: integer, format: int64 }
...
    FengYuCatalogEntry:
      type: object
      description: Field names mirror the FengYu host's MarketplaceCatalogEntry contract.
      properties:
        id: { type: string }
        name: { type: string }
        description: { type: string }
        version: { type: string }
        author: { type: string }
        icon: { type: string, format: uri, nullable: true }
        category: { type: string }
        permissions: { type: array, items: { type: string } }
        homepage: { type: string, nullable: true }
        downloadUrl: { type: string, description: Direct HMAC-ticketed package URL }
        official: { type: boolean }
        sha256: { type: string }
        signature: { type: string, nullable: true }
        keyId: { type: string, nullable: true }